Address final code review feedback
