+++ benchmark_test.go
@@ -2,6 +2,7 @@ package strings2
 
 import (
 	"testing"
+	"unicode"
 )
 
 func BenchmarkUpperCaseFirst_ASCII_Short(b *testing.B) {
@@ -179,5 +180,21 @@ func BenchmarkSplitMixCase(b *testing.B) {
 	b.ReportAllocs()
 	for i := 0; i < b.N; i++ {
 		_, _ = WordsToFormattedCase(words, OptionMixCaseSupport(), OptionDelimiter("-"))
+	}
+}
+
+func BenchmarkPerformCaseFirst(b *testing.B) {
+	s := "test"
+	fn := unicode.ToUpper
+	for i := 0; i < b.N; i++ {
+		performCaseFirst(s, fn)
+	}
+}
+
+func BenchmarkPerformCaseFirst_Long(b *testing.B) {
+	s := "teststringwithmorecharacters"
+	fn := unicode.ToUpper
+	for i := 0; i < b.N; i++ {
+		performCaseFirst(s, fn)
 	}
 }
